+++ AGENTS.md
@@ -1,21 +1,22 @@
 # AGENTS.md
 
-このファイルはAIエージェント向けの作業指示です。
+このファイルはAIエージェント向けの作業指示です。人間向けの詳しい説明、セットアップ、運用手順は原則 `README.md` に書いてください。
 
-- `AGENTS.md`: AIが間違えないための短い設計指示、作業ルール、判断基準を書く
-- `README.md`: 人間が理解・運用するための詳しい説明、セットアップ、運用手順を書く
+## 役割分担
 
-人間向けの詳しい説明を追加したい場合は、原則 `README.md` に書いてください。
-設計判断・実装時の注意・AIに守ってほしいルールは `AGENTS.md` に書いてください。
+- `AGENTS.md`: AIが実装時に守る設計方針、作業ルール、判断基準、間違えやすい仕様を書く。
+- `README.md`: アプリ概要、開発手順、環境変数、起動手順、デプロイ手順、画面説明、トラブルシューティングを書く。
+- 技術スタック、ディレクトリ構成、URL仕様、検索条件データ構造、API方針を変えた場合は、`AGENTS.md` と `README.md` の更新要否を確認する。
+- 小さなUI調整や一時的な実験だけなら、ドキュメント更新は必須ではない。
 
 ## 作業ルール
 
 - `AGENTS.md` だけを根拠にせず、変更前に必ず該当ファイルと周辺の呼び出し元を読む。
-- 既存の設計・命名・TypeScript/JavaScript混在方針・Sass構成・Store/API層の使い方に合わせる。
+- 既存の設計、命名、TypeScript/JavaScript混在方針、Sass/SCSS構成、Store/API層の使い方に合わせる。
 - ユーザーの未コミット変更を勝手に戻さない。
 - 生成物や依存関係の大きな更新は、明確に必要な場合だけ行う。
-- 技術スタック、ディレクトリ構成、URL仕様、検索条件データ構造、API方針を変えた場合は、`AGENTS.md` と `README.md` の更新要否を確認する。
-- 小さなUI調整や一時的な実験だけなら、ドキュメント更新は必須ではない。
+- 既存JSファイルをTSへ変換する場合は、呼び出し元、Webpack設定、型定義への影響を確認する。
+- 依存を移動・追加・削除する場合は、実行時依存かビルド時依存かを確認し、`package.json` と `package-lock.json` を揃える。
 
 ## 技術スタック
 
@@ -29,12 +30,11 @@
 | バックエンド連携 | TogoVar API                                    |
 | パッケージ       | npm                                            |
 
-補足:
-
 - Node.js は `22.x` 前提。`.nvmrc` と `package.json` の `engines.node` を確認する。
-- `tsconfig.json` は `strict: true` だが `allowJs: true`。既存JSとTSが共存しているため、周辺ファイルの粒度に合わせて変更する。
+- `tsconfig.json` は `strict: true` かつ `allowJs: true`。既存JSとTSが共存しているため、周辺ファイルの粒度に合わせて変更する。
+- `app/frontend/server/` は `tsconfig.server.json` と `npm run typecheck:server` で確認する。
 
-## 重要なディレクトリ方針
+## ディレクトリ方針
 
 ```txt
 app/frontend/
@@ -47,10 +47,10 @@ app/frontend/
     components/  UIコンポーネント・画面部品
       AdvancedSearch/   Advanced Search全体制御
       Condition/        条件入力UI・クエリ変換
-        ConditionDiseaseSearch/             疾患検索Lit要素
+        ConditionDiseaseSearch/                  疾患検索Lit要素
         ConditionPathogenicityPredictionSearch/  病原性予測検索Lit要素
-        ConditionValueEditor/               条件値入力UI
-        queryBuilders/                      条件→APIクエリ変換
+        ConditionValueEditor/                    条件値入力UI
+        queryBuilders/                           条件→APIクエリ変換
       Karyotype/        染色体ビジュアライズ
       PanelView/        サイドバーパネル
       Results/          検索結果テーブル
@@ -62,12 +62,12 @@ app/frontend/
   views/         Pugテンプレート
 ```
 
-- UIクラス・Web Componentsは `app/frontend/src/components/` に置く（Lit使用の有無で分けない）。
+- UIクラス・Web Componentsは `app/frontend/src/components/` に置く。Lit使用の有無では分けない。
 - Advanced Search 本体の制御ファイルは `app/frontend/src/components/AdvancedSearch/` に置く。
 - StoreやURL反映など、アプリ状態に関わる処理は `app/frontend/src/store/` に置く。
 - DOM生成の小さな共通処理は `app/frontend/src/utils/dom/` の既存ヘルパーを優先する。
 - 参照ゲノム別の検索条件マスタは `app/frontend/assets/GRCh37` / `app/frontend/assets/GRCh38` を確認する。
-- PugテンプレートやSassの構成を変える場合は、関連する `app/frontend/views/` と `stylesheets/` の両方を確認する。
+- PugテンプレートやSass/SCSSの構成を変える場合は、関連する `app/frontend/views/` と `app/frontend/stylesheets/` の両方を確認する。
 
 ## デプロイ / 配信方針
 
@@ -75,8 +75,7 @@ app/frontend/
 - GitHub Pages向けの `.github/workflows/publish.yml` も `npm run build` 後に `dist/` を公開するだけで、`app/frontend/server/` は起動しない。
 - `app/frontend/server/` はローカル開発サーバー、または `npm run start:prod` でビルド済みファイルをローカル確認するための補助サーバーとして扱う。
 - 本番Dockerで `node app/frontend/server` / `npm run start:prod` を起動している証跡は、現時点で共有された設定からは見つかっていない。
-- ただし顧客側の実デプロイ設定がこのリポジトリ外にある可能性があるため、`app/frontend/server/`、`npm start`、`npm run start:prod`、`express` 関連依存を削除する前に、Dockerfile / Compose / nginx設定 / 起動コマンドを確認する。
-- `logger.ts` と `ip` 依存はフロントエンド用Expressサーバーの起動ログ用途。Expressサーバー本体より削除リスクは低いが、削除時は `npm start` / `npm run start:prod` の動作も合わせて整理する。
+- 顧客側の実デプロイ設定がリポジトリ外にある可能性があるため、`app/frontend/server/`、`npm start`、`npm run start:prod`、`express` 関連依存を削除する前に、Dockerfile / Compose / nginx設定 / 起動コマンドを確認する。
 
 ## Advanced Search 方針
 
@@ -89,9 +88,9 @@ app/frontend/
   - `components/Condition/ConditionValues.ts` と `components/Condition/ConditionValueEditor/*`: 条件入力UI
 - 条件の検索クエリ化は `components/Condition/queryBuilders/` に集約する。
 - Advanced Search のURL共有は `?mode=advanced&q=<Base64 JSON>` を使う。
-  - encode/decode処理は `app/frontend/src/store/advancedSearchURL.ts` に集約する。
-  - URLから復元した条件をViewへ戻す処理は `components/AdvancedSearch/AdvancedSearchConditionRestorer.ts` を確認する。
-  - Base64はURL中で壊れやすいため、生成時は `encodeURIComponent` を通す。
+- encode/decode処理は `app/frontend/src/store/advancedSearchURL.ts` に集約する。
+- URLから復元した条件をViewへ戻す処理は `components/AdvancedSearch/AdvancedSearchConditionRestorer.ts` を確認する。
+- Base64はURL中で壊れやすいため、生成時は `encodeURIComponent` を通す。
 - `setAdvancedSearchCondition()` は検索条件をStoreへ保存し、URLへ反映し、検索を実行する。呼び出しタイミングを増やす場合は二重検索に注意する。
 - 条件UIを変更した場合は、検索クエリ、URL共有、URL復元の3点が同じ構造を扱えるか確認する。
 
@@ -102,7 +101,7 @@ app/frontend/
 - API通信は `app/frontend/src/api/fetchData.ts` など既存API層に合わせる。
 - コンポーネントやViewから直接URLやfetchの仕様を増やす前に、既存のStore/API層へ寄せられるか確認する。
 - 戻る/進む、URL貼り付け、モード切り替えでは、Store更新と検索実行が重複しやすい。変更時は初期表示・タブ切替・履歴操作を分けて考える。
-- popstate ハンドラ内での `storeManager.setData('searchMode', ...)` は `pushState` を発火させてしまう。`setSearchModeFromHistory(mode)` を使うことで、popstate 中の pushState を防ぐ（StoreManager の `#fromHistory` フラグで制御する）。
+- popstate ハンドラ内での `storeManager.setData('searchMode', ...)` は `pushState` を発火させてしまう。`setSearchModeFromHistory(mode)` を使うことで、popstate 中の pushState を防ぐ。
 
 ## TypeScript / JavaScript 方針
 
@@ -121,14 +120,15 @@ app/frontend/
 - UI状態は、可能なら既存の `data-*` 属性や状態クラスを利用する。
 - セレクターを追加する前に、既存の `object/component` / `object/project` に同じ責務のスタイルがないか確認する。
 - `app/frontend/stylesheets/foundation/_reset.scss` は**変更禁止**。Josh W. Comeau's CSS Reset の定義をそのまま維持する。
-- `tr` に `border` を使うテーブルは、必ずそのテーブルのセレクター自身に `border-collapse: collapse` を明示する（`_reset.scss` に書けないため各コンポーネントで対応）。
+- `tr` に `border` を使うテーブルは、必ずそのテーブルのセレクター自身に `border-collapse: collapse` を明示する。`_reset.scss` には書けないため各コンポーネントで対応する。
 
 ## コメント規約
 
 - 新規・更新するコメントは日本語を基本にする。
 - コメントは「何をしているか」だけでなく、「なぜその形にしているか」を優先する。
 - 単にコードを読み上げるコメントは書かない。
-- 全ての公開・非公開の関数、メソッド、getter/setter には必ず「WHY（なぜその形にしているか）」を1行目に日本語で書く。必要なら2行目以降で補足してよいが、1行目のWHYは省略しない。
+- 全ての公開・非公開の関数、メソッド、getter/setter には必ず「WHY（なぜその形にしているか）」を1行目に日本語で書く。
+- 必要なら2行目以降で補足してよいが、1行目のWHYは省略しない。
 - リファクタリング時にコメントが古くなったら、コードと一緒に必ず更新する。
 - 既存に英語コメントが残っている場合、触った範囲では日本語化を検討する。ただし大規模なコメント翻訳だけの差分は避ける。
 
@@ -149,27 +149,6 @@ npm run lint:css
 npm run typecheck
 ```
 
-注意:
-
 - このリポジトリには `npm run check` は定義されていない。
 - 環境によって `node` / `npm` がPATHに無い場合がある。その場合は実行できなかったことを報告する。
 - Rails/バックエンド側を触った場合は、READMEのBackend手順と既存のRubyコマンドを確認する。
-
-## READMEとの分担
-
-READMEに書くもの:
-
-- アプリ概要
-- ローカル開発手順
-- 環境変数
-- バックエンド/フロントエンドの起動手順
-- デプロイ手順
-- 人間向けの詳しい画面説明
-- トラブルシューティング
-
-AGENTSに書くもの:
-
-- AIが実装時に守るべき設計方針
-- 変更時の判断基準
-- 間違えやすいデータ名・URL仕様・API方針
-- コード配置ルール
